refactor(table): css import paths
