+++ packages/table/src/index.ts
@@ -1,4 +1,6 @@
-export { default } from './react/index'
+import './css/index.css'
+import Table from './react/index'
+export default Table
 export * from './react/sentinel'
 export * from './react/sticky'
