+++ packages/table/src/react/__stories__/drawer.story.tsx
@@ -1,7 +1,7 @@
 import { Meta, Story } from '@storybook/react/types-6-0'
 import React from 'react'
 
-import Table from '../index'
+import Table from '../../index'
 import { generateUser } from './seed'
 import { ExpandButton, OutlineBox } from './shared'
