+++ packages/table/styles.css
@@ -1,0 +1,1 @@
+@import "./dist/esm/css/index.css" ;
